yta_backend/pyproject.toml
[project]
name = "yta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "praw>=7.8.1",
    "python-decouple>=3.8",
]
